update: avoid using Reflection to get Performance each time
