+++ src/clocks/monotonic/mod.rs
@@ -3,10 +3,16 @@ mod windows;
 #[cfg(target_os = "windows")]
 pub use self::windows::Monotonic;
 
-#[cfg(target_arch = "wasm32")]
-mod wasm;
-#[cfg(target_arch = "wasm32")]
-pub use self::wasm::Monotonic;
+#[cfg(all(target_arch = "wasm32", target_os = "unknown"))]
+mod wasm_browser;
+#[cfg(all(target_arch = "wasm32", target_os = "unknown"))]
+pub use self::wasm_browser::Monotonic;
+
+#[cfg(all(target_arch = "wasm32", target_os = "wasi"))]
+mod wasm_wasi;
+#[cfg(all(target_arch = "wasm32", target_os = "wasi"))]
+pub use self::wasm_wasi::Monotonic;
+
 
 #[cfg(not(any(target_os = "windows", target_arch = "wasm32")))]
 mod unix;
